RAG/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
pandas
sentence-transformers
faiss-cpu
groq
python-dotenv
pydantic